+++ components/frontend/src/share/Share.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { Header } from 'semantic-ui-react';
+import { PermLinkButton } from '../widgets/Button';
+
+export function Share({ title, url }) {
+    return (
+        <>
+            <Header size="small">
+                {title}
+            </Header>
+            <PermLinkButton url={url} />
+        </>
+    )
+}