+++ components/frontend/src/subject/SubjectTitle.js
@@ -1,15 +1,14 @@
 import React, { useContext } from 'react';
-import { Grid, Header, Icon, Menu, Tab } from 'semantic-ui-react';
-import { Comment } from '../fields/Comment';
-import { StringInput } from '../fields/StringInput';
-import { SubjectType } from './SubjectType';
+import { Header, Icon, Menu, Tab } from 'semantic-ui-react';
 import { HeaderWithDetails } from '../widgets/HeaderWithDetails';
 import { DeleteButton, ReorderButtonGroup } from '../widgets/Button';
 import { ChangeLog } from '../changelog/ChangeLog';
+import { Share } from '../share/Share';
 import { delete_subject, set_subject_attribute } from '../api/subject';
 import { DataModel } from '../context/DataModel';
 import { EDIT_REPORT_PERMISSION, ReadOnlyOrEditable } from '../context/Permissions';
 import { FocusableTab } from '../widgets/FocusableTab';
+import { SubjectParameters } from './SubjectParameters';
 
 function SubjectHeader({ subject_type }) {
     return (
@@ -21,51 +20,6 @@ function SubjectHeader({ subject_type }) {
                 </Header.Subheader>
             </Header.Content>
         </Header>
-    )
-}
-
-function SubjectConfiguration({ subject, subject_uuid, subject_name, reload }) {
-    return (
-        <Grid stackable>
-            <Grid.Row columns={3}>
-                <Grid.Column>
-                    <SubjectType
-                        id={`${subject_uuid}-type`}
-                        set_value={(value) => set_subject_attribute(subject_uuid, "type", value, reload)}
-                        subject_type={subject.type}
-                    />
-                </Grid.Column>
-                <Grid.Column>
-                    <StringInput
-                        id={`${subject_uuid}-title`}
-                        requiredPermissions={[EDIT_REPORT_PERMISSION]}
-                        label="Subject title"
-                        placeholder={subject_name}
-                        set_value={(value) => set_subject_attribute(subject_uuid, "name", value, reload)}
-                        value={subject.name}
-                    />
-                </Grid.Column>
-                <Grid.Column>
-                    <StringInput
-                        id={`${subject_uuid}-subtitle`}
-                        label="Subject subtitle"
-                        requiredPermissions={[EDIT_REPORT_PERMISSION]}
-                        set_value={(value) => set_subject_attribute(subject_uuid, "subtitle", value, reload)}
-                        value={subject.subtitle}
-                    />
-                </Grid.Column>
-            </Grid.Row>
-            <Grid.Row>
-                <Grid.Column>
-                    <Comment
-                        id={`${subject_uuid}-comment`}
-                        requiredPermissions={[EDIT_REPORT_PERMISSION]}
-                        set_value={(value) => set_subject_attribute(subject_uuid, "comment", value, reload)}
-                        value={subject.comment}
-                    />
-                </Grid.Column>
-            </Grid.Row>
-        </Grid>
     )
 }
 
@@ -86,9 +40,20 @@ export function SubjectTitle({ report, subject, subject_uuid, first_subject, las
     const dataModel = useContext(DataModel)
     const current_subject_type = dataModel.subjects[subject.type] || { name: "Unknown subject type" };
     const subject_name = subject.name || current_subject_type.name;
+    const subjectUrl = `${window.location}#${subject_uuid}`
     const panes = [
-        { menuItem: <Menu.Item key="configuration"><Icon name="settings" /><FocusableTab>{"Configuration"}</FocusableTab></Menu.Item>, render: () => <Tab.Pane><SubjectConfiguration subject={subject} subject_uuid={subject_uuid} subject_name={subject_name} reload={reload} /></Tab.Pane> },
-        { menuItem: <Menu.Item key="changelog"><Icon name="history" /><FocusableTab>{"Changelog"}</FocusableTab></Menu.Item>, render: () => <Tab.Pane><ChangeLog subject_uuid={subject_uuid} timestamp={report.timestamp} /></Tab.Pane> }
+        {
+            menuItem: <Menu.Item key="configuration"><Icon name="settings" /><FocusableTab>{"Configuration"}</FocusableTab></Menu.Item>,
+            render: () => <Tab.Pane><SubjectParameters subject={subject} subject_uuid={subject_uuid} subject_name={subject_name} reload={reload} /></Tab.Pane>
+        },
+        {
+            menuItem: <Menu.Item key="changelog"><Icon name="history" /><FocusableTab>{"Changelog"}</FocusableTab></Menu.Item>,
+            render: () => <Tab.Pane><ChangeLog subject_uuid={subject_uuid} timestamp={report.timestamp} /></Tab.Pane>
+        },
+        {
+            menuItem: <Menu.Item key="share"><Icon name="share square" /><FocusableTab>{'Share'}</FocusableTab></Menu.Item>,
+            render: () => <Tab.Pane><Share title="Subject permanent link" url={subjectUrl} /></Tab.Pane>
+        }
     ];
     return (
         <HeaderWithDetails level="h2" header={subject_name} subheader={subject.subtitle} style={{ marginTop: 50 }}>